Created all necessary views
Created the views needed.
Added buttons to admin panel.
Created the route for GradeController.
Errors might appear when merging the work with controllers, because of inclaritiy of exact content. Only keep the display_view or redirect()->to() paths.
Languages entries might not be exact or created. To fix when merging.