Link subjects and modules to a domain
Updated how subjects and modules are linked to a domain. Forms have been updated to no longer let the user choose the parent (course plans for domains, domains for subjects) of the entry.
All is made in the details of a course, where clicking on the button to add a new domain will automatically get his parent accordingly to the course plan we currently are, and clicking the button to add a new subject will automatically get the domain in which the button is.

Linking modules to a domain is made on a new page, where we can select all modules we want to link to the domain.

Created two common views (in \orif\plafor\Views): for pages titles and form validation errors.

Created the views for teaching domains title CRUD.