Access permissions review
All methods permissions have been updated, to correspond to the revised acces permissions document.

Added functions in access permission helper, to verify that a specified user is an apprentice, trainer, or admin, and a function to get the access level of a given user.

Comments have been reworked, notably to match the new permissions.
Also, they now use the generic manage_entry view.

Warning : a database field has been renamed (in the comment table) in this commit. Be sure to migrate the changes or manually update the database field.