UI changes in progressView component
Changed the display of the ProgressView component to enhance the UX. (Can be viewed in the view of an apprentice.)

The button to develop the operational competences has been remade to be bigger.
The link to access the acquisition status of the objectives of an operational competence is now on its name instead of its symbol.

Refactored a bit the progressView.js file.