Updated the forms for managing course plans
Updated the course_plan, competence_domains, operational_competences, objectives forms to delete the parent element dropdown option and replace it by a hidden id field. The parent element name is displayed to the user.

When linking modules to a domain, the parent domain where we link modules to is now shown.

Updated views documentation in consequence.